Update index.js
I wanted to do the exercise with switch(), because it looked to me like that it is easier,  searched a little bit, and found this as a solution. The code runs in the console, but it does not on Udemy. Does someone think that this can be one of the solutions to the exercise? Thank you.